move localization tests back to their own file
