+++ packages/react-composites/tests/browser/call/Localization.test.ts-snapshots/localized-call-configuration-page-chromium.png
